Ported last of DDL tests
